requirements.txt
python-pptx==0.6.22
pandas